Added tracer
